+++ packages/e2e.generated.catchall/src/generatedNoCheck/index.ts
@@ -15,7 +15,16 @@ export {
   editOsdkTestObject,
 } from './ontology/actions.js';
 export * as $Actions from './ontology/actions.js';
-export { Athlete, bus_1, FooInterface, InterfaceNoProps, OsdkTestInterface } from './ontology/interfaces.js';
+export {
+  Athlete,
+  bus_1,
+  CollateralConcernCandidate,
+  CollateralConcernList,
+  EsongInterfaceA,
+  FooInterface,
+  InterfaceNoProps,
+  OsdkTestInterface,
+} from './ontology/interfaces.js';
 export * as $Interfaces from './ontology/interfaces.js';
 export {
   BgaoNflPlayer,
@@ -24,6 +33,8 @@ export {
   Country_1,
   DherlihyComplexObject,
   Employee,
+  EsongIssues,
+  EsongPds,
   FintrafficAis,
   GtfsTripTrackObject,
   MatthewvsDevOrderEmbedding,