+++ packages/e2e.generated.catchall/src/generatedNoCheck/ontology/interfaces.ts
@@ -1,5 +1,8 @@
 export { Athlete } from './interfaces/Athlete.js';
 export { bus_1 } from './interfaces/bus_1.js';
+export { CollateralConcernCandidate } from './interfaces/CollateralConcernCandidate.js';
+export { CollateralConcernList } from './interfaces/CollateralConcernList.js';
+export { EsongInterfaceA } from './interfaces/EsongInterfaceA.js';
 export { FooInterface } from './interfaces/FooInterface.js';
 export { InterfaceNoProps } from './interfaces/InterfaceNoProps.js';
 export { OsdkTestInterface } from './interfaces/OsdkTestInterface.js';