+++ packages/create-app.template.tutorial-todo-aip-app/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @osdk/create-app.template.tutorial-todo-aip-app
 
+## 2.4.0-beta.13
+
+## 2.4.0-beta.12
+
+## 2.4.0-beta.11
+
+## 2.4.0-beta.10
+
 ## 2.4.0-beta.9
 
 ## 2.4.0-beta.8