+++ packages/example-generator/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @osdk/example-generator
 
+## 0.12.0-beta.13
+
+### Patch Changes
+
+- @osdk/create-app@2.4.0-beta.13
+- @osdk/create-widget@3.1.0-beta.2
+
+## 0.12.0-beta.12
+
+### Patch Changes
+
+- @osdk/create-app@2.4.0-beta.12
+- @osdk/create-widget@3.1.0-beta.2
+
+## 0.12.0-beta.11
+
+### Patch Changes
+
+- @osdk/create-app@2.4.0-beta.11
+- @osdk/create-widget@3.1.0-beta.2
+
+## 0.12.0-beta.10
+
+### Patch Changes
+
+- @osdk/create-widget@3.1.0-beta.2
+- @osdk/create-app@2.4.0-beta.10
+
 ## 0.12.0-beta.9
 
 ### Patch Changes